requirement.txt
annotated-types==0.6.0
asgiref==3.7.2
colorama==0.4.6
Django==4.2.6
django-ninja==1.0b2
orjson==3.9.10
packaging==23.2
pip-review==1.3.0
PLTable==1.0.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
